+++ NewStore.Domain/Entities/Users/User.cs
@@ -9,7 +9,7 @@ namespace NewStore.Domain.Entities.Users
         public bool IsActive { get; set; } = true;
         public string? Lastname { get; set; }
         public string Email { get; set; }
-        public int? Age { get; set; }
+        public short? Age { get; set; }
         public string Password { get; set; }
         public ICollection<UserInRole> UserInRoles { get; set; }
     }
